aura_mediabuyer/requirements.txt
facebook-business>=19.0.0
python-dotenv>=1.0.0 